pyproject.toml
[tool.poetry]
name = "crypto-trading-system"
version = "0.1.0"
description = "Automated Crypto Day Trading System using FastAPI - A robust API for managing automated crypto trading operations with Binance integration"
authors = ["James Fincher"]
readme = "README.md"
packages = [{include = "."}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlalchemy = "^2.0.23"
pydantic = {extras = ["email"], version = "^2.5.1"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
binance-connector = "^3.5.1"
pandas = "^2.1.3"
numpy = "^1.26.2"
httpx = "^0.25.2"
python-dateutil = "^2.8.2"
aiosqlite = "^0.19.0"
email-validator = "^2.2.0"
alembic = "^1.13.0"
bcrypt = "^4.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=. --cov-report=term-missing"
